# proto/service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/service.proto",
    "version": "version not set"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/example/call/{name}": {
      "get": {
        "operationId": "ApiCall",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/hbchenResponse"
            }
          }
        },
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "services",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "ExampleService"
        ]
      }
    }
  },
  "definitions": {
    "ResponseChain": {
      "type": "object",
      "properties": {
        "service": {
          "$ref": "#/definitions/hbchenService"
        },
        "ctx": {
          "type": "string"
        },
        "chain": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ResponseChain"
          }
        }
      }
    },
    "hbchenNode": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "host": {
          "type": "string"
        }
      }
    },
...
    "hbchenResponse": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string"
        },
        "chain": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ResponseChain"
          }
        }
      }
    },
    "hbchenService": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "node": {
          "$ref": "#/definitions/hbchenNode"
        },
        "services": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/hbchenService"
          }
        }
      }
    }
  }
}